--- packages/datacommons-api/src/datacommons/api/app.py
@@ -1,29 +1,0 @@
-# Copyright 2025 Google LLC.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from fastapi import FastAPI
-
-from datacommons.api.endpoints.routers import node_router
-
-# FastAPI initialization
-app = FastAPI(
-    title="Data Commons API",
-    version="0.1.0",
-    openapi_url="/openapi.json",
-    docs_url="/docs",
-    redoc_url="/redoc",
-    debug=True,
-)
-
-app.include_router(node_router.router, tags=["nodes"])
