+++ packages/datacommons-db/pyproject.toml
@@ -1,18 +1,12 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
 [project]
 name = "datacommons-db"
 dynamic = ["version"]
 description = 'Data Commons DB'
+license = "Apache-2.0"
 readme = "README.md"
-requires-python = ">=3.11"
-license = "MIT"
+requires-python = ">=3.12,<3.13"
 keywords = []
-authors = [
-  { name = "Dan Noble", email = "dwnoble@google.com" },
-]
+authors = []
 classifiers = [
   "Development Status :: 4 - Beta",
   "Programming Language :: Python",
@@ -32,12 +26,9 @@ Documentation = "https://github.com/datacommonsorg/datacommons#readme"
 Issues = "https://github.com/datacommonsorg/datacommons/issues"
 Source = "https://github.com/datacommonsorg/datacommons"
 
-[tool.hatch.build.targets.wheel]
-packages = ["src/datacommons"]
+[build-system]
+requires = ["uv", "setuptools"]
+build-backend = "setuptools.build_meta"
 
-[tool.hatch.metadata]
-allow-direct-references = true
-
-[tool.hatch.version]
-path = "src/datacommons/db/__init__.py"
-
+[tool.setuptools.dynamic]
+version = {attr = "datacommons_db.version.__version__"}
