--- packages/datacommons-root/src/datacommons/__init__.py
@@ -1,1 +1,0 @@
-version = "0.1.0"
