+++ packages/datacommons-schema/datacommons_schema/models/jsonld.py
